pyproject.toml
[project]
name = "ampav-parakeet"
version = "0.0.12"
description = "Parakeet tooling for the AMPAV environment"
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",   
]
authors = [
    {name = "Brian Wheeler", email="bdwheele@iu.edu"},
]

requires-python = ">=3.12"
dependencies = [
    "ampav-core>=0.0.12",
    "nemo_toolkit[asr]==2.7.2",
]

[project.urls]
Homepage = "https://github.com/AMPAV/ampav-parakeet"
Repository = "https://github.com/AMPAV/ampav-parakeet.git"
Issues = "https://github.com/AMPAV/ampav-parakeet/issues"


[project.scripts]
amapv_parkeet_transcribe = "ampav.parakeet.transcribe:cli_parakeet_transcribe"


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"